# backend/openapi/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "description" : "Contrato HTTP do CRM PNP. Tenant é sempre derivado da credencial.",
    "title" : "CRM PNP API",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "paths" : {
...
    "/api/canais/{id}" : {
      "put" : {
        "operationId" : "atualizar_4",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CanalRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CanalResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Requisição ou JSON inválido"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Autenticação ausente ou expirada"
          },
          "403" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Acesso negado"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Recurso não encontrado"
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Conflito de estado ou idempotência"
          },
          "422" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Regra de negócio ou validação semântica"
          },
          "429" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Limite de requisições excedido"
          },
          "500" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Falha interna com detalhe seguro"
          }
        },
        "tags" : [ "canal-controller" ]
      }
    },
...
    "/api/contatos/{id}" : {
      "delete" : {
        "operationId" : "excluir_2",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Requisição ou JSON inválido"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Autenticação ausente ou expirada"
          },
          "403" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Acesso negado"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Recurso não encontrado"
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Conflito de estado ou idempotência"
          },
          "422" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Regra de negócio ou validação semântica"
          },
          "429" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Limite de requisições excedido"
          },
          "500" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Falha interna com detalhe seguro"
          }
        },
        "tags" : [ "contact-controller" ]
      },
      "get" : {
        "operationId" : "obter",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ContatoResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Requisição ou JSON inválido"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Autenticação ausente ou expirada"
          },
          "403" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Acesso negado"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Recurso não encontrado"
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Conflito de estado ou idempotência"
          },
          "422" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Regra de negócio ou validação semântica"
          },
          "429" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Limite de requisições excedido"
          },
          "500" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Falha interna com detalhe seguro"
          }
        },
        "tags" : [ "contact-controller" ]
      },
      "put" : {
        "operationId" : "atualizar_3",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ContatoRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ContatoResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Requisição ou JSON inválido"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Autenticação ausente ou expirada"
          },
          "403" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Acesso negado"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Recurso não encontrado"
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Conflito de estado ou idempotência"
          },
          "422" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Regra de negócio ou validação semântica"
          },
          "429" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Limite de requisições excedido"
          },
          "500" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Falha interna com detalhe seguro"
          }
        },
        "tags" : [ "contact-controller" ]
      }
    },
...
    "/api/integracoes/http" : {
      "get" : {
        "operationId" : "listar_1",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Resumo"
                  }
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Requisição ou JSON inválido"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Autenticação ausente ou expirada"
          },
          "403" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Acesso negado"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Recurso não encontrado"
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Conflito de estado ou idempotência"
          },
          "422" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Regra de negócio ou validação semântica"
          },
          "429" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Limite de requisições excedido"
          },
          "500" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Falha interna com detalhe seguro"
          }
        },
        "tags" : [ "http-connector-controller" ]
      },
      "post" : {
        "operationId" : "criar_2",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ConnectorRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Resumo"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Requisição ou JSON inválido"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Autenticação ausente ou expirada"
          },
          "403" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Acesso negado"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Recurso não encontrado"
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Conflito de estado ou idempotência"
          },
          "422" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Regra de negócio ou validação semântica"
          },
          "429" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Limite de requisições excedido"
          },
          "500" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Falha interna com detalhe seguro"
          }
        },
        "tags" : [ "http-connector-controller" ]
      }
    },
    "/api/integracoes/http/{id}" : {
      "put" : {
        "operationId" : "atualizar_2",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ConnectorRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Resumo"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Requisição ou JSON inválido"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Autenticação ausente ou expirada"
          },
          "403" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Acesso negado"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Recurso não encontrado"
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Conflito de estado ou idempotência"
          },
          "422" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Regra de negócio ou validação semântica"
          },
          "429" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Limite de requisições excedido"
          },
          "500" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Falha interna com detalhe seguro"
          }
        },
        "tags" : [ "http-connector-controller" ]
      }
    },
    "/api/integracoes/http/{id}/ativar" : {
      "post" : {
        "operationId" : "ativar",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Resumo"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Requisição ou JSON inválido"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Autenticação ausente ou expirada"
          },
          "403" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Acesso negado"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Recurso não encontrado"
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Conflito de estado ou idempotência"
          },
          "422" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Regra de negócio ou validação semântica"
          },
          "429" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Limite de requisições excedido"
          },
          "500" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Falha interna com detalhe seguro"
          }
        },
        "tags" : [ "http-connector-controller" ]
      }
    },
    "/api/integracoes/http/{id}/pausar" : {
      "post" : {
        "operationId" : "pausar",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Resumo"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Requisição ou JSON inválido"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Autenticação ausente ou expirada"
          },
          "403" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Acesso negado"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Recurso não encontrado"
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Conflito de estado ou idempotência"
          },
          "422" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Regra de negócio ou validação semântica"
          },
          "429" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Limite de requisições excedido"
          },
          "500" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Falha interna com detalhe seguro"
          }
        },
        "tags" : [ "http-connector-controller" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "CanalRequest" : {
        "type" : "object",
        "properties" : {
          "identificadorExterno" : {
            "type" : "string",
            "maxLength" : 120,
            "minLength" : 0
          },
          "nome" : {
            "type" : "string",
            "maxLength" : 120,
            "minLength" : 0
          },
          "segredoWebhook" : {
            "type" : "string",
            "maxLength" : 300,
            "minLength" : 0
          },
          "tipo" : {
            "type" : "string",
            "enum" : [ "LIVE_CHAT", "TELEGRAM", "WHATSAPP_CLOUD", "WHATSAPP_EVOLUTION", "INSTAGRAM" ]
          },
          "token" : {
            "type" : "string",
            "maxLength" : 300,
            "minLength" : 0
          }
        },
        "required" : [ "nome", "tipo" ]
      },
      "CanalResponse" : {
        "type" : "object",
        "properties" : {
          "ativo" : {
            "type" : "boolean"
          },
          "estadoRemoto" : {
            "type" : "string"
          },
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "identificadorExterno" : {
            "type" : "string"
          },
          "nome" : {
            "type" : "string"
          },
          "pendenciasRemotas" : {
            "type" : "integer",
            "format" : "int32"
          },
          "temSegredoWebhook" : {
            "type" : "boolean"
          },
          "temToken" : {
            "type" : "boolean"
          },
          "tipo" : {
            "type" : "string"
          },
          "ultimaFalhaRemotaEm" : {
            "type" : "string",
            "format" : "date-time"
          },
          "ultimaReconciliacaoEm" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "ErroCampo" : {
        "properties" : {
          "campo" : {
            "type" : "string"
          },
          "mensagem" : {
            "type" : "string"
          }
        }
      },
      "ErroResponse" : {
        "properties" : {
          "campos" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ErroCampo"
            }
          },
          "codigo" : {
            "type" : "string"
          },
          "correlacaoId" : {
            "type" : "string"
          },
          "detail" : {
            "type" : "string"
          },
          "instance" : {
            "type" : "string",
            "format" : "uri"
          },
          "momento" : {
            "type" : "string",
            "format" : "date-time"
          },
          "status" : {
            "type" : "integer",
            "format" : "int32"
          },
          "title" : {
            "type" : "string"
          },
          "type" : {
            "type" : "string",
            "format" : "uri"
          }
        }
      },
...
    }
  }
}